fix for class construct/destruct
MEMO:
__del__() is called not when exiting the scope of a method but when ref count is 0.
