+++ aws/terraform/shared_modules/aws_s3_bucket_core/test/pytest/test_main.py
@@ -1,11 +1,9 @@
 import logging
 import os
 import uuid
-from functools import singledispatchmethod
 
 import boto3
 from tf_pytest.aws import AwsIAMPolicyTester
-from typing_extensions import override
 
 import pytest
 
@@ -14,33 +12,14 @@ _logger = logging.getLogger(__name__)
 _logger.setLevel(os.environ.get("LOG_LEVEL", "INFO"))
 
 
-@pytest.fixture(scope="session")
-def delete_all_object(init_destroy):
-    yield
-
-    _logger.info("delete all object")
-
-    root = init_destroy
-    aws_s3_bucket = root.module.default.aws_s3_bucket.default
-
-    bucket_name = aws_s3_bucket.values["bucket"]
-    s3 = boto3.resource("s3").Bucket(bucket_name).objects.all().delete()
-
-
 class AwsIAMPolicyTesterS3Write(AwsIAMPolicyTester):
     def __init__(self, aws_iam_role_arn: str):
         super().__init__(aws_iam_role_arn)
 
-    def __del__(self):
+    def close(self):
         pass
 
-    @singledispatchmethod
-    @override
-    def test(self):
-        raise NotImplementedError(f"This {type(self).__name__} does not implement test().")
-
-    @test.register
-    def _test_impl(self, bucket_name: str, object_key: str) -> bool:
+    def test(self, bucket_name: str, object_key: str) -> bool:
         try:
             s3 = self._session.client("s3")
             s3.put_object(Bucket=bucket_name, Key=object_key, Body="test")
@@ -74,17 +53,16 @@ class AwsIAMPolicyTesterS3Write(AwsIAMPolicyTester):
 def test_write_object(tfstate_skip_apply, delete_all_object, request, iam_role, expect):
     root = tfstate_skip_apply
 
-    aws_s3_bucket = root.module.default.aws_s3_bucket.default
     aws_iam_role = root.aws_iam_role.defaults[iam_role]
+    aws_s3_bucket = root.module.default.aws_s3_bucket.default
 
     object_key = str(uuid.uuid4())
 
-    tester = AwsIAMPolicyTesterS3Write(**{"aws_iam_role_arn": aws_iam_role.values["arn"]})
+    with AwsIAMPolicyTesterS3Write(**{"aws_iam_role_arn": aws_iam_role.values["arn"]}) as tester:
+        assert tester.test(aws_s3_bucket.values["bucket"], object_key) == expect
 
-    assert tester.test(aws_s3_bucket.values["bucket"], object_key) == expect
 
-
-class AwsIAMPoliyTesterS3Read(AwsIAMPolicyTester):
+class AwsIAMPolicyTesterS3Read(AwsIAMPolicyTester):
     def __init__(self, aws_iam_role_arn: str, aws_s3_bucket_name: str, aws_s3_object_key: str):
         super().__init__(aws_iam_role_arn)
 
@@ -92,10 +70,14 @@ class AwsIAMPoliyTesterS3Read(AwsIAMPolicyTester):
         self._aws_s3_object_key = aws_s3_object_key
 
         client = boto3.client("s3")
+
+        _logger.info(f"create object: s3://{self._aws_s3_bucket_name}/{self._aws_s3_object_key}")
         client.put_object(Bucket=self._aws_s3_bucket_name, Key=self._aws_s3_object_key, Body="test")
 
-    def __del__(self):
+    def close(self):
         client = boto3.client("s3")
+
+        _logger.info(f"delete object: s3://{self._aws_s3_bucket_name}/{self._aws_s3_object_key}")
         client.delete_object(Bucket=self._aws_s3_bucket_name, Key=self._aws_s3_object_key)
 
     def test(self):
@@ -131,17 +113,16 @@ class AwsIAMPoliyTesterS3Read(AwsIAMPolicyTester):
 def test_read_object(tfstate_skip_apply, delete_all_object, request, iam_role, expect):
     root = tfstate_skip_apply
 
+    aws_iam_role = root.aws_iam_role.defaults[iam_role]
+    aws_s3_bucket = root.module.default.aws_s3_bucket.default
+
     object_key = str(uuid.uuid4())
 
-    aws_s3_bucket = root.module.default.aws_s3_bucket.default
-    aws_iam_role = root.aws_iam_role.defaults[iam_role]
-
-    tester = AwsIAMPoliyTesterS3Read(
+    with AwsIAMPolicyTesterS3Read(
         **{
             "aws_iam_role_arn": aws_iam_role.values["arn"],
             "aws_s3_bucket_name": aws_s3_bucket.values["bucket"],
             "aws_s3_object_key": object_key,
         }
-    )
-
-    assert tester.test() == expect
+    ) as tester:
+        assert tester.test() == expect
